Create start
I imported the necessary namespace, 'System,' which provides access to essential functions and classes in C#.

I created a class called 'Program' that contains the main logic of the game. The 'Main' method is the entry point, where the program execution begins.

To generate random numbers, I instantiated a 'Random' object called 'random_number_generator.' This object will help me create a random number between 0 and 20.

I displayed a welcoming message to the player and asked for their name, storing it in the 'userName' variable.

Afterwards, I displayed the game instructions, informing the player to guess a number between 0 and 20 and providing the option to terminate the program by entering -1.

To keep track of the number of guesses and store the player's guess, I initialized 'number_of_guesses' and 'player_guess' variables.

I set up a 'while' loop that runs until the player decides to terminate the game.

Inside the loop, I prompt the player for their guess, read their input, and check if they want to quit the game by entering -1.

If the player continues playing, I compare their guess to the random number and provide feedback, either suggesting a higher or lower guess.

Overall, I've created an interactive game that allows users to test their guessing skills while providing a friendly user interface.